+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Release date: ``2019-xx-xx``
 - `NXPY-80 <https://jira.nuxeo.com/browse/NXPY-80>`__: Stick with pytest < 4 to prevent internal error due to the use of deprecated ``pytest_namespace``
 - `NXPY-81 <https://jira.nuxeo.com/browse/NXPY-81>`__: Fix flake8 errors and add flake8 to the CI
 - `NXPY-82 <https://jira.nuxeo.com/browse/NXPY-82>`__: Fix ``test_convert_xpath()``
+- `NXPY-83 <https://jira.nuxeo.com/browse/NXPY-83>`__: Fix ``test_convert()`` and ``test_convert_given_converter()``
 
 2.0.4
 -----