+++ CHANGELOG.rst
@@ -11,6 +11,15 @@ Release date: ``2019-xx-xx``
 - `NXPY-82 <https://jira.nuxeo.com/browse/NXPY-82>`__: Fix ``test_convert_xpath()``
 - `NXPY-83 <https://jira.nuxeo.com/browse/NXPY-83>`__: Fix ``test_convert()`` and ``test_convert_given_converter()``
 - `NXPY-84 <https://jira.nuxeo.com/browse/NXPY-84>`__: Handle ``list`` type in operation parameters
+- `NXPY-87 <https://jira.nuxeo.com/browse/NXPY-87>`__: Add an upload helper to control the chunk uploads
+
+Technical changes
+-----------------
+
+- Added ``Batch.get_uplaoder()``
+- Added nuxeo/uploads.py::\ ``API.get_uploader()``
+- Added ``callback`` keyword argument to nuxeo/uploads.py::\ ``API.upload()``
+- Added nuxeo/uploads.py::\ ``Uploader``
 
 2.0.4
 -----